updated carrousel.js name, made autoslide, and text over image
