+++ client/src/App.css
@@ -16,6 +16,32 @@
   min-height: calc(100vh - 400px);
 }
 
+#overlap {
+  position: relative;
+  color: #000;
+  width:750px
+}
+
+#caption {
+  top: 50%;
+  left: auto;
+  position: absolute;
+}
+
+#quote-container {
+  display: flex;
+  justify-content: center;
+  top: 50%;
+  left: 0%;
+  z-index: 999;
+}
+
+.daily-quote-image {
+  opacity: 0.5;
+  max-height: 490px;
+  min-width: 100%;
+}
+
 /* FOOTER STYLING */
 
 /* MEDIA QUERY */
