+++ client/src/pages/Landing.js
@@ -2,7 +2,7 @@ import React from "react";
 
 import Auth from "../utils/auth";
 
-import ControlledCarousel from "../components/ControlledCarousel";
+import Carousel from "../components/Carousel";
 import SearchLocation from "./SearchLoc";
 
 const Landing = () => {
@@ -15,7 +15,11 @@ const Landing = () => {
           </>
         ) : (
           <>
-            <ControlledCarousel />
+          <div id="overlap">
+          <h1>We've helped these hangry people</h1>
+            <Carousel />
+          <h2>lets help you too</h2>
+          </div>
           </>
         )}
       </div>
